Implementation of the ZFILESYSTEM feature. A ZFILESYSTEM is a standard Eucalyptus/OpenStack instance filesystem, that has been compressed. Supported compression schemes are: gzip, bzip2, compress, lzma
The underlying operating system must possess the appropriate command-line utility,
in order to successfully de-compress a provided image. ImageProxy will log a failure
(but continue to operate) if a desired utility is missing.

A ZFILESYSTEM can be specified in place of a FILESYSTEM in the ImageProxy metafile.
The signature to be used for a ZFILESYSTEM is an shasum of the *compressed* filesystem (*NOT*
an shasum of the uncompressed filesystem used to create the ZFILESYSTEM).