Integrate OpenStack support into ImageProxy.
OpenStack uses euca2ools, just as Eucalyptus does.
OpenStack does things slightly differently, however, so the Eucalyptus
upload-and-register script needs to be altered slightly.

The script now loops on the output of "euca-describe-images" and
awaits the image's transition into "available" state.
Under Eucalyptus, this is "instantaneous", upon registration.
Under OpenStack, we have to wait on a number of state transitions, before
the image finally becomes "available."

As a result of the requirement to loop on image availability, a
new property has been introduced, imageproxy.registrationTimeout.
This timeout allows the site administrator to specify a maximum time
for which to wait upon the registered image to become available.

If this property is not specified, the default is 1800 seconds.
This value was decided based on Paul's statement that large images
could take "10s of minutes" to become "available" under OpenStack.