Add tar archived files to ZFILESYSTEM support. Paul has discovered that large images can be packaged as sparse files; this often occurs if the image has a lot of unused space.
gzip does not retain the "sparseness" of a file; neither do most other
compression programs.

tar does, however, maintain "sparseness."
The nature of tar archives as a "container" format requires that
we have a convention for the naming of the filesystem within the
tar archive.

The filesystem should be a file named "filesystem" within the tar archive.
Terribly clever, I know. ;)